Update databaseBackup.php
Changed to accommodate MySQL PDO connection and also changed the queries to contain backticks for table names. This is important so that the database connects using PDO (like the rest of the application on this version) and so that the table name 'groups' is differentiated from the MySQL reserved keyword.